+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### New features
 
+- .NET 8.0 has been added as a target framework.
+
 ### Changes to existing features
 
 ### Bugs fixed in this release